requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2
djangorestframework-simplejwt
django-filter
djoser
pytest
pytest-django
pytest-pythonpath
